Increase scale_p0/scale_p1 by 32 times
This corresponds to a change in the `fft_shift` settings in the
dibas.conf file on the BTL side.  Having these two (three if you count
two here) related settings in two places is not ideal.  Probably one
should move to be with the other at some point (but maybe there's a
reason they can't be together?)

The `fft_shift` settings were increased to avoid overflows in the
FFT/PFB that would result in splattering the RFI over multiple channels
across the band.  The necessitated a corresponding increase in these
scaling values.